Compiler-warning cleanup and clang-tidy pass
Mechanical warning cleanup across the tree plus clang-tidy findings (header
guards, override specifiers, unused-parameter marks, MSVC-only pragma
guards). Upstream PR #736 covers the C side; the C++ batch is the planned
follow-up.

+++ src/actors/yoshi_egg/render.inc.c
@@ -20,7 +20,7 @@ void render_actor_yoshi_egg(Camera* camera, Mat4 arg1, struct YoshiValleyEgg* eg
     f32 temp_f0;
 
     size_t actorIdx = CM_FindActorIndex((struct Actor*) egg);
-    if (-1 == actorIdx) {
+    if (-1 == (s32) actorIdx) {
         printf("[render_actor_yoshi_egg] Could not find actor for FI, skipping\n");
         return;
     }
