+++ src/code_80086E70.c
@@ -1311,7 +1311,7 @@ void func_8008A1D0(s32 objectIndex, s32 cameraId, s32 arg2, s32 arg3) {
     clear_object_flag(objectIndex, 0x00100000 | VISIBLE);
     temp_v0 = get_horizontal_distance_to_camera(objectIndex, camera);
     if (CVarGetInteger("gNoCulling", 0) == 1) {
-        temp_v0 = MIN(temp_v0, arg3 * arg3);
+        temp_v0 = MIN(temp_v0, (u32) (arg3 * arg3));
     }
     if (temp_v0 < 0x2711U) {
         var_a2 = 0x5555;
@@ -1357,9 +1357,9 @@ s32 func_8008A364(s32 objectIndex, s32 cameraId, u16 arg2, s32 arg3) {
     clear_object_flag(objectIndex, 0x00020000 | VISIBLE);
     dist = get_horizontal_distance_to_camera(objectIndex, camera);
     if (CVarGetInteger("gNoCulling", 0) == 1) {
-        dist = MIN(dist, (arg3 * arg3) - 1);
+        dist = MIN(dist, (u32) ((arg3 * arg3) - 1));
     }
-    if (dist < (arg3 * arg3)) {
+    if (dist < (u32) (arg3 * arg3)) {
         set_object_flag(objectIndex, 0x00020000);
         if (dist < 0x2711U) {
             var_a2 = 0x5555;
